Merge Debugprobe and Picoports
Luckily OpenOCD does not detect CMSIS-DAP debuggers by their vendor and
product ID, but by the product and interface strings, which we can
freely chose. The vendor and product ID are fixed because the dln2
driver is matched using them.

For this merge we switch the debugprobe build to a library build. Then
Debugprobe can be linked together with Picoports. Debugprobe already
provides a CDC ACM / UART interfaces, so we can drop the previous UART
interface implementation. Debugprobe requires some modifications to work
alongside Picoports. These are documented in commits of the submodule
branch.

Debugprobe uses FreeRTOS. Picoports was not written with a preemtive
kernel in mind, so as a start we're letting the Picoports loop run in
the FreeRTOS idle task with the scheduler suspended. While initial tests
showed that this works, it might influence timing too much and make the
debugger unstable, time will tell.

We're ignoring Debugprobe's usb_description.c and instead adapt
Picoports' usb_description.c to incude a second vendor interface for
Debugprobe's DAP interface.